pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openai-billing-monitor"
version = "0.1.0"
description = "A Python library for monitoring and controlling OpenAI API costs"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "tiktoken>=0.4.0",
]

[project.optional-dependencies]
gui = ["tkinter"]
dev = ["pytest", "black", "flake8", "mypy"]

[project.scripts]
openai-billing-gui = "openai_billing.gui.main:main"

[project.urls]
Homepage = "https://github.com/lemonhall/openai-billing-monitor"
Repository = "https://github.com/lemonhall/openai-billing-monitor"
Issues = "https://github.com/lemonhall/openai-billing-monitor/issues"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
